pyproject.toml
[project]
name = "alxhttp"
version = "0.9.94"
authors = [{ name = "Alexander Macdonald", email = "alex@alexmac.cc" }]
description = "A better base server for AioHTTP"
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]
dependencies = [
  "asyncpg-stubs>=0.30.2",
  "basedpyright>=1.31.4",
  "pip>=25.0.1",
  "pytest-cov>=6.0.0",
  "pytest-mock>=3.14.0",
  "pytest>=8.3.4",
  'aiohttp ~= 3.9',
  'asyncpg ~= 0.29',
  'pglast ~= 7.7',
  'pydantic ~= 2.7',
  'pyhumps ~= 3.8',
  'redis ~= 6.4',
  'watchdog ~=6.0',
  "twine>=6.2.0",
  "ty>=0.0.2",
]

[project.urls]
Homepage = "https://github.com/alexmac/alxhttp"

[project.optional-dependencies]
xray = ['aws-xray-sdk >= 2.13']

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = [
  "alxhttp",
  "alxhttp.middleware",
  "alxhttp.pydantic",
  "alxhttp.tests",
  "alxhttp.typescript",
  "alxhttp.typescript.wrappers",
]

[tool.ruff]
line-length = 200
indent-width = 2

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true

[tool.isort]
profile = "black"
indent = "  "
line_length = 200

[tool.pytest.ini_options]
log_file_level = "INFO"
log_file = "output/pytest/logs.txt"
log_file_format = "%(asctime)s [%(levelname)8s] [%(name)s] %(message)s (%(filename)s:%(lineno)s)"
minversion = "6.0"
addopts = "-vv --cov=alxhttp --cov-report=html --cov-report=term --cov-report=xml"
testpaths = ["tests", "integration"]
filterwarnings = ["ignore:.*`json_encoders` is deprecated.*"]
cache_dir = "output/pytest/cache"

[tool.coverage.run]
branch = true
data_file = "output/coverage/.coverage"

[tool.coverage.html]
directory = "output/coverage"

[tool.coverage.xml]
output = "output/coverage/coverage.xml"

[tool.pyright]
ignore = ["output"]
reportAny = false
reportExplicitAny = false
reportUnusedCallResult = false # I want to enabled this, but the IO "write" method hiyts this case
reportConstantRedefinition = false # reports glopbals as always constant, not sure i agree
reportUnknownVariableType = false # too noisy when walking dicts from json.loads
reportUnknownMemberType = false # too noisy when walking dicts from json.loads
reportPrivateUsage = false
reportUnusedParameter = false
